+++ packages/services/assetlibraryhistory/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@cdf/assetlibrary-history-v2.2.3](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/compare/@cdf/assetlibrary-history-v2.2.2...@cdf/assetlibrary-history-v2.2.3) (2019-10-22)
+
+
+### Bug Fixes
+
+* **assetlibrary-history:** fixing broken tests ([142d4f3](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/commit/142d4f3bc00637d6a7c1ec3a157522dbd0245ea5))
+
+# [@cdf/assetlibrary-history-v2.2.2](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/compare/@cdf/assetlibrary-history-v2.2.1...@cdf/assetlibrary-history-v2.2.2) (2019-10-21)
+
+
+### Bug Fixes
+
+* **assetlibrary-history:** fixing broken tests ([f0ff307](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/commit/f0ff307cea631cf90e041a92be820aeccb4bff95))
+
+# [@cdf/assetlibrary-history-v2.2.1](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/compare/@cdf/assetlibrary-history-v2.2.0...@cdf/assetlibrary-history-v2.2.1) (2019-10-21)
+
+
+### Bug Fixes
+
+* **assetlibraryhistory:** updated to match v2 assetlibrary models ([e6254ee](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/commit/e6254eee67adbe05d48a0c18467dd2d63550b981))
+
 # [@cdf/assetlibrary-history-v2.2.0](https://git-codecommit.us-west-2.amazonaws.com/v1/repos/cdf-core/compare/@cdf/assetlibrary-history-v2.1.5...@cdf/assetlibrary-history-v2.2.0) (2019-06-30)